pyproject.toml
[tool.poetry]
name = "article-reader-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Yeasin Mollik <yeasin1603@cseku.ac.bd>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.11"
export-to-telegraph = "^0.0.143"
urlextract = "^1.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
